chore: split codebase into seperate files
