+++ src/utils/cmd.ts
@@ -1,0 +1,12 @@
+import { exec } from 'child_process';
+
+export const cmd = (commandLine: string) => {
+  return new Promise<string>((resolve, reject) => {
+    exec(commandLine, (error, stdout, stderr) => {
+      if (error) {
+        reject(error);
+      }
+      resolve(stdout || stderr);
+    });
+  });
+};
